Added 1ms timer interrupt to handle encoder unwrapping. Every 1ms, the TIM counter is checked for possible overflow, if the current counts is over a configurable threshold, the counter is reset, and the current counts are stored in a global static int32_t variable. mock_enc.c was updated to reflect the change in the API due to counts now being a int32_t variable.

+++ tst/src/mock_enc.c
@@ -1,7 +1,7 @@
 #include "enc.h"
 
-uint16_t gs_enc_cnt;
+int32_t gs_enc_cnt;
 
-uint16_t ENC_GetCnt(void) {
+int32_t ENC_GetCnt(void) {
   return gs_enc_cnt;
 }
